+++ apps/species/urls.py
@@ -1,9 +1,10 @@
 from django.urls import path
 
-from .views import YearlistView
+from .views import SpeciesView, YearlistView
 
 app_name = "species"
 
 urlpatterns = [
+    path("", SpeciesView.as_view(), name="list"),
     path("yearlist/<int:year>/", YearlistView.as_view(), name="yearlist"),
 ]