+++ CHANGELOG.md
@@ -8,6 +8,7 @@ a concise history of changes made.
 ## Latest
 - Added form fields for editing translated text in the Django Admin.
 - Added retries for eBird API calls and page loads.
+- Added the Species page for showing year lists, site list, personal lists, etc.
 
 ## 2025.05.31
 - Added notifications to the Latest News page.